Remove PARSER_PRIVATE() macro
Accessing the private data using the G_TYPE_INSTANCE_GET_PRIVATE()
macro is neither free nor good practice. The macro performs a type
check and other operations, like locking/unlocking mutexes in a
thread-enabled application, that make this kind of access a lot
slower than it seems.

The best practice in this case consists of using a pointer to the
private data structure inside the instance structure.

This patch removes the PARSER_PRIVATE() macro and all of its uses.